requirements.txt
stripe
twilio
google-auth
google-auth-oauthlib
google-api-python-client
requests
